deleteThing: updated

+++ backend/controllers/stuff.js
@@ -1,4 +1,5 @@
 const Thing = require("../models/Things");
+const fs = require("fs");
 
 exports.createThing = (req, res, next) => {
   const thingObject = JSON.parse(req.body.thing);
@@ -38,6 +39,7 @@ exports.modifyThing = (req, res, next) => {
       if (thing.userId != req.auth.userId) {
         res.status(401).json({ message: "Non autorisé" });
       } else {
+        console.log(thing);
         Thing.updateOne(
           { _id: req.params.id },
           { ...thingObject, _id: req.params.id }
@@ -52,9 +54,25 @@ exports.modifyThing = (req, res, next) => {
 };
 
 exports.deleteThing = (req, res, next) => {
-  Thing.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({ message: "Objet supprimé !" }))
-    .catch((error) => res.status(400).json({ error }));
+  Thing.findOne({ _id: req.params.id })
+    .then((thing) => {
+      console.log(thing);
+      if (thing.userId != req.auth.userId) {
+        res.status(401).json({ message: "Non autorisé" });
+      } else {
+        const filename = thing.imageUrl.split("/images/")[1];
+        fs.unlink(`images/${filename}`, () => {
+          Thing.deleteOne({ _id: req.params.id })
+            .then(() => {
+              res.status(200).json({ message: "Objet supprimé !" });
+            })
+            .catch((error) => res.status(401).json({ error }));
+        });
+      }
+    })
+    .catch((error) => {
+      res.status(500).json({ error });
+    });
 };
 
 exports.getOneThing = (req, res, next) => {
